Add functions for finding min and max values
